refactor: convert is_distinct_from operators to functions
Following PR review feedback, convert IS DISTINCT FROM and IS NOT DISTINCT FROM from operators to functions.

Changes:
- Remove Ash.Query.Operator.DistinctFrom and NotDistinctFrom
- Add Ash.Query.Function.IsDistinctFrom and IsNotDistinctFrom
- Move optimization logic to new() instead of simplify()
- Update test descriptions to reflect eager evaluation behavior

Rationale:
- Functions cannot be overloaded, which is appropriate here since is_distinct_from only needs [:any, :any] signature
- Optimization (converting to != or == when NULL is impossible) happens at creation time in new()
- simplify() is only for filter comparison context, not needed for these functions
