+++ lib/ash/query/function/is_distinct_from.ex
@@ -2,37 +2,34 @@
 #
 # SPDX-License-Identifier: MIT
 
-defmodule Ash.Query.Operator.DistinctFrom do
+defmodule Ash.Query.Function.IsDistinctFrom do
   @moduledoc """
-  left is_distinct_from right
+  is_distinct_from(left, right)
 
   SQL's IS DISTINCT FROM operator.
   Unlike `!=`, this operator treats NULL as a comparable value.
 
   When both sides cannot return NULL, this simplifies to `!=` for better performance.
   """
-  use Ash.Query.Operator,
-    operator: :is_distinct_from,
-    name: :distinct_from,
-    predicate?: true,
-    types: [:same, :any]
+  use Ash.Query.Function, name: :is_distinct_from, predicate?: true
 
-  @impl Ash.Query.Operator
-  def new(left, right) do
+  def args, do: [[:any, :any]]
+
+  def returns, do: [:boolean]
+
+  def evaluate_nil_inputs?, do: true
+
+  def new([left, right]) do
     if Ash.Expr.can_return_nil?(left) || Ash.Expr.can_return_nil?(right) do
-      {:ok, struct(__MODULE__, left: left, right: right)}
+      {:ok, struct(__MODULE__, arguments: [left, right])}
     else
       Ash.Query.Operator.new(Ash.Query.Operator.NotEq, left, right)
     end
   end
 
-  @impl Ash.Query.Operator
-  def evaluate(%{left: left, right: right}) do
+  def evaluate(%{arguments: [left, right]}) do
     {:known, Comp.not_equal?(left, right)}
   end
-
-  @impl Ash.Query.Operator
-  def evaluate_nil_inputs?, do: true
 
   def can_return_nil?(_), do: false
 end
